@@ -4,7 +4,7 @@
     "license": "SEE LICENSE IN LICENSE.txt",
     "repository": {
         "type": "hg",
-        "url": "https://bitbucket.org/viperproject/viper-ide"
+        "url": "https://github.com/viperproject/viper-ide"
     },
     "version": "0.0.1",
     "publisher": "viper-admin",
@@ -14,23 +14,24 @@
     },
     "dependencies": {
         "decompress-zip": "^0.3.1",
-        "find-java-home": "github:aterga/node-find-java-home",
         "glob-to-regexp": "^0.4.0",
+        "locate-java-home": "git+https://github.com/viperproject/locate-java-home.git",
         "mkdirp": "^0.5.1",
-        "portfinder": "^1.0.17",
+        "node-ipc": "^9.1.4",
         "request": "^2.88.2",
         "sound-play": "^1.1.0",
         "stream-json": "^1.7.1",
         "sudo-prompt": "8.2.3",
         "tree-kill": "^1.2.2",
-        "vscode-languageserver": "^5.1.0",
-        "vscode-uri": "^1.0.8"
+        "vscode-languageserver": "^5.2.1",
+        "vscode-uri": "^2.1.1"
     },
     "devDependencies": {
-        "@types/node": "^10.14.6",
+        "@types/mocha": "^8.0.3",
+        "@types/node": "^14.11.8",
         "@types/stream-json": "^1.5.1",
-        "typescript": "^3.4.5",
-        "vscode": "^1.1.33"
+        "@types/vscode": "^1.43.0",
+        "typescript": "^4.0.3"
     },
     "scripts": {
         "compile": "tsc -p .",
