@@ -1,22 +1,57 @@
-// 
-// PLEASE DO NOT MODIFY / DELETE UNLESS YOU KNOW WHAT YOU ARE DOING  
+// this file is taken from the helloworld-test-sample from https://github.com/microsoft/vscode-extension-samples
+
+// Copyright (c) Microsoft Corporation
 //
-// This file is providing the test runner to use when running extension tests.
-// By default the test runner in use is Mocha based.
-// 
-// You can provide your own test runner if you want to override it by exporting
-// a function run(testRoot: string, clb: (error:Error) => void) that the extension
-// host can call to run the tests. The test runner is expected to use console.log
-// to report the results back to the caller. When the tests are finished, return
-// a possible error to the callback or null if none.
+// All rights reserved. 
+//
+// MIT License
+//
+// Permission is hereby granted, free of charge, to any person obtaining a copy of this software and associated documentation 
+// files (the "Software"), to deal in the Software without restriction, including without limitation the rights to use, copy,
+// modify, merge, publish, distribute, sublicense, and/or sell copies of the Software, and to permit persons to whom the Software 
+// is furnished to do so, subject to the following conditions:
+//
+// The above copyright notice and this permission notice shall be included in all copies or substantial portions of the Software.
+//
+// THE SOFTWARE IS PROVIDED *AS IS*, WITHOUT WARRANTY OF ANY KIND, EXPRESS OR IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES
+// OF MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT HOLDERS 
+// BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT 
+// OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
 
-var testRunner = require('vscode/lib/testrunner');
+import * as path from 'path';
+import * as Mocha from 'mocha';
+import * as glob from 'glob';
 
-// You can directly control Mocha options by uncommenting the following lines
-// See https://github.com/mochajs/mocha/wiki/Using-mocha-programmatically#set-options for more info
-testRunner.configure({
-	ui: 'bdd', 		// the TDD UI is being used in extension.test.ts (suite, test, etc.)
-	useColors: true // colored output from test results
-});
+export async function run(): Promise<void> {
+	// Create the mocha test
+	const mocha = new Mocha({
+		ui: 'tdd',
+		// Installing and starting Viper might take some minutes
+        timeout: 600_000, // ms
+        color: true,
+	});
 
-module.exports = testRunner;
+	const testsRoot = path.resolve(__dirname, '..');
+
+	const files: Array<string> = await new Promise((resolve, reject) =>
+        glob(
+            "**/*.test.js",
+            {
+                cwd: testsRoot,
+            },
+            (err, result) => {
+                if (err) reject(err)
+                else resolve(result)
+            }
+        )
+    )
+
+	// Add files to the test suite
+    files.forEach(f => mocha.addFile(path.resolve(testsRoot, f)));
+
+    const failures: number = await new Promise(resolve => mocha.run(resolve));
+
+    if (failures > 0) {
+        throw new Error(`${failures} tests failed.`)
+    }
+}
